Added argparse options
Added positive and negative variables ARGPARSE and NOARGPARSE. Use the positive, because the negative may go away in the future.
if ARGPARSE or if not ARGPARSE
Added options inside an "if ARGPARSE", and assigned the dictionary "args" if not. I would rather it be transparent, so I may create a class with the variables.