Changes for future mysql query support
MagentoCtl.db_cache_table took in part of the globalconfig dict that was larger than was needed. As a result, the function referenced two levels deeper, but needed nothing in between. This was so the update depth and query depth matched.

There is now a MysqlCtl class that includes the code to build a query line. It is important to note that it is NOT sanitized so fuzzing local.xml could potentially cause bad things to happen.